cleanup code and move github functions to SharedFunctions
